Convert from exporting a full express app to just a Router instance
Initialising a new app expects the views directory to be set, and defaults it otherwise. This means that routing through a new app will always destroy any parent app's setting of a views directory causing requests to fail.

Exporting a router instead of an app means that req.app still points at the parent app and so any settings are persisted correctly.