Return relative paths from backlink middleware
In order to be able to support configured back links that link out of the scope of a baseUrl we need to be able to distinguish between relative and absolutely defined links, so that it can be determined later if they need to be prefixed with req.baseUrl.

A back link which is explicitly configured as an absolute path will remain an absolute path, but all dynamically determined back links will be relative.