Set controller error class in the constructor
At the moment the error class is only set as part of the wizard constructor, which means that controllers will only have their error class set when in the context of a wizard, and not when constructed standalone. This means that testing error handling and generation cannot be done whilst exncapsualting the controller.

Setting the error class on a controller at construction time rather than in the wizard post-construction allows the behaviour to be inherited in standalone controllers.