Add slash to backlinks.
I have no idea why the code removes the backlinks if the backlink is just relying on the last step.

It introduces bugs for us. All other tests expect the backlinks to have slashes so I can't figure out why.

This PR re-introduces slashes for backlinks that are just checked from the session steps.

The one usecase I can see where backlinks would need to remove the slash is where the last step in the flow was just '/'

In that case this PR checks for the existance for a single referring '/' and sets it to '' which will disable the backlink.