Extend backlinks whitelisting to include the full forms journey history
Where a set of backLinks are whitelisted, the back-links controller checks to see if the list includes the preceding page in the form steps or, if not, the referring page. However, there may be scenarios whereby a previous step further back in the jouney should be whitelisted. This update handles this by adding another check to see if any steps - from this or any other wizard instance - have been whitelisted. There is a use case in the Passports renewals service whereby mulitple form wizard instances are used and the journey can naturally flow between them.