Remove use of express session is no session defined.
This is very dangerous. If you can't access redis (or your store of choice) for some reason (it's down for example) then the wizard will make a session anyway using in memory sessions which are known to be leaky leaky leaky.

It should error at this point which the developer can capture and deal with.