bugfix - fix for infinite loop
On address lookup step in a multi journey form, it is possible to get into an infinite loop if the fork target is a step that has already been traversed. This is because the top level array of steps was not passed between recursive calls to check step has not been traversed.

* pass allSteps array in getAllPossibleSteps
* check step not already in allSteps before traversing