Fix typo in usage example
The usage example is incorrect for the default next step for step3. This
commit fixes the example shown to provide a default string for the 'next'
property on step3. So if the participant is old enough, i.e they have
not been sent to the 'not-old-enough' step they will be sent to step4.